Adds a game of guess the number to the project.
Inside of this guessing game, the user is given the ability to set the
highest value the computer can give for a number between 0 -
user_number. After inserting a number they are given a game against the
computer where they must guess what number the computer came up with in
the background.