Updates the README with work flow and future projs.
Rock Paper Scissors, Address Book, and Text Based Adventure have been
all added to this of ideas that will be done in this collection of
projects. The ideas section contain the stuff that will be produced later. The doing stuff is what is in progress. The complete stuff is the completed projects and the stuck is what is recorded to stuck on.